# CSharp/REST/RestTodoList-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TODO LIST API",
    "description": "exemple d'API REST avec de la documentation",
    "contact": {
      "name": "Fred BERTON",
      "email": "frederic.Berton@capgemini.com"
    },
    "license": {
      "name": "LGPL"
    },
    "version": "1.00"
  },
  "paths": {
    "/api/v1/TodoList": {
      "get": {
        "tags": [
          "TodoList"
        ],
        "operationId": "GetTodoItems",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/TodoList/GetTodoItem/{id}": {
      "get": {
        "tags": [
          "TodoList"
        ],
        "operationId": "GetTodoItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/TodoList/CreateTodoItem": {
      "get": {
        "tags": [
          "TodoList"
        ],
        "operationId": "CreateTodoItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateTodoItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateTodoItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateTodoItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/TodoList/UpdateTodoItem/{id}": {
      "put": {
        "tags": [
          "TodoList"
        ],
        "operationId": "UpdateTodoItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateTodoItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateTodoItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateTodoItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/TodoList/DeleteTodoItem/{id}": {
      "delete": {
        "tags": [
          "TodoList"
        ],
        "operationId": "DeleteTodoItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateOrUpdateTodoItem": {
        "type": "object",
        "properties": {
          "titre": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TodoItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "titre": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}